+++ test/pkgs/sample/PkgWithTestCompatOutsideWorkspace/src/PkgWithTestCompatOutsideWorkspace.jl
@@ -1,0 +1,1 @@
+module PkgWithTestCompatOutsideWorkspace end